fix: audio/screenshare isn't stopped when joining breakouts
There are two scenarios where media in the main room isn't properly
stopped when joining breakouts:
  - Audio: when manually requesting to join a breakout the user was not
    invited to
  - Screen share: when accepting an invite to a breakout room.
The media stop code for breakouts is also split between the two modes
(invitation and manual join), which contributes to inconsistencies.

This commit rewrites and centralizes the procedure to stop the user's media
in the main room in a hook called useStopMediaOnMainRoom. Both the
join-confirmation and manual join procedure now use that hook, which
guarantees audio, screen sharing and video to be stopped when necessary.
