+++ bigbluebutton-html5/imports/ui/components/breakout-room/breakout-room/component.tsx
@@ -18,12 +18,7 @@ import { BREAKOUT_ROOM_END_ALL, BREAKOUT_ROOM_REQUEST_JOIN_URL, USER_TRANSFER_VO
 import useMeeting from '/imports/ui/core/hooks/useMeeting';
 import TimeRemaingPanel from './components/timeRemaining';
 import BreakoutMessageForm from './components/messageForm';
-import {
-  finishScreenShare,
-  forceExitAudio,
-  stopVideo,
-} from './service';
-import { useExitVideo, useStreams } from '/imports/ui/components/video-provider/hooks';
+import { useStopMediaOnMainRoom } from '/imports/ui/components/breakout-room/hooks';
 import useDeduplicatedSubscription from '/imports/ui/core/hooks/useDeduplicatedSubscription';
 
 interface BreakoutRoomProps {
@@ -115,6 +110,7 @@ const BreakoutRoom: React.FC<BreakoutRoomProps> = ({
   const [breakoutRoomEndAll] = useMutation(BREAKOUT_ROOM_END_ALL);
   const [breakoutRoomTransfer] = useMutation(USER_TRANSFER_VOICE_TO_MEETING);
   const [breakoutRoomRequestJoinURL] = useMutation(BREAKOUT_ROOM_REQUEST_JOIN_URL);
+  const stopMediaOnMainRoom = useStopMediaOnMainRoom();
 
   const layoutContextDispatch = layoutDispatch();
   const isRTL = layoutSelect((i: Layout) => i.isRTL);
@@ -156,12 +152,10 @@ const BreakoutRoom: React.FC<BreakoutRoomProps> = ({
       if (breakout && breakout.joinURL) {
         window.open(breakout.joinURL, '_blank');
         setRequestedBreakoutRoomId('');
+        stopMediaOnMainRoom(presenter);
       }
     }
-  }, [breakouts]);
-
-  const exitVideo = useExitVideo();
-  const streams = useStreams();
+  }, [breakouts, stopMediaOnMainRoom, presenter]);
 
   return (
     <Styled.Panel
@@ -245,15 +239,7 @@ const BreakoutRoom: React.FC<BreakoutRoomProps> = ({
                                   requestJoinURL(breakout.breakoutRoomId);
                                 } else {
                                   window.open(breakout.joinURL, '_blank');
-                                  // leave main room's audio,
-                                  // and stops video and screenshare when joining a breakout room
-                                  forceExitAudio();
-                                  stopVideo(exitVideo, streams);
-                                  logger.info({
-                                    logCode: 'breakoutroom_join',
-                                    extraInfo: { logType: 'user_action' },
-                                  }, 'joining breakout room closed audio in the main room');
-                                  if (presenter) finishScreenShare();
+                                  stopMediaOnMainRoom(presenter);
                                 }
                               }}
                               disabled={requestedBreakoutRoomId}
